# Persistent wait thread