--- a/NEWS.md
+++ b/NEWS.md
@@ -1,7 +1,12 @@
 # nanonext 1.3.0.9004 (development)
 
+#### Behavioural Change
+
+* User interrupts while using `call_aio_()`, `collect_aio_()` or the equivalent `[]` method now cancel the 'aio' operation causing the Aio to return an 'errorValue 20 | Operation Canceled'.
+
 #### Updates
 
+* Performs interruptible 'aio' waits using a single dedicated thread, rather than launching new threads, for higher performance and efficiency.
 * Performance enhancements for 'ncurlAio' and 'recvAio' promises methods.
 * Updates bundled 'libnng' to v1.9.0 stable release.
 * The package has a shiny new hex logo.