--- a/R/utils.R
+++ b/R/utils.R
@@ -356,6 +356,7 @@ serial_config <- function(class, sfunc, ufunc, vec = FALSE)
 #' if (Sys.info()[["sysname"]] == "Linux") {
 #'   rm(list = ls())
 #'   gc()
+#'   .Call(nanonext:::rnng_thread_shutdown)
 #'   Sys.sleep(1L)
 #'   .Call(nanonext:::rnng_fini)
 #' }