--- a/man/zzz.Rd
+++ b/man/zzz.Rd
@@ -14,6 +14,7 @@ to run the examples.
 if (Sys.info()[["sysname"]] == "Linux") {
   rm(list = ls())
   gc()
+  .Call(nanonext:::rnng_thread_shutdown)
   Sys.sleep(1L)
   .Call(nanonext:::rnng_fini)
 }