--- a/src/thread.c
+++ b/src/thread.c
@@ -23,16 +23,17 @@
 // threads callable and messenger ----------------------------------------------
 
 int nano_wait_thread_created = 0;
-
 nng_thread *nano_wait_thr;
+nng_aio *nano_shared_aio;
+
 nng_mtx *nano_wait_mtx;
 nng_cv *nano_wait_cv;
 int nano_wait_condition = 0;
 
-nng_mtx *nano_shared_mtx;
-nng_cv *nano_shared_cv;
-static int nano_shared_condition = 0;
-nng_aio *nano_shared_aio;
+static void check_interrupt(void * data) {
+  (void) data;
+  R_CheckUserInterrupt();
+}
 
 // # nocov start
 // tested interactively
@@ -225,16 +226,15 @@ static void rnng_wait_thread(void *args) {
       nng_mtx_unlock(nano_wait_mtx);
       break;
     }
-    nano_wait_condition = 0;
     nng_mtx_unlock(nano_wait_mtx);
 
     nng_aio_wait(nano_shared_aio);
 
-    nng_mtx_lock(nano_shared_mtx);
+    nng_mtx_lock(nano_wait_mtx);
     nano_shared_aio = NULL;
-    nano_shared_condition = 1;
-    nng_cv_wake(nano_shared_cv);
-    nng_mtx_unlock(nano_shared_mtx);
+    nano_wait_condition = 0;
+    nng_cv_wake(nano_wait_cv);
+    nng_mtx_unlock(nano_wait_mtx);
   }
 
 }
@@ -276,11 +276,6 @@ static void thread_disp_finalizer(SEXP xptr) {
 
 }
 
-static void check_interrupt(void * data) {
-  (void) data;
-  R_CheckUserInterrupt();
-}
-
 SEXP rnng_thread_shutdown(void) {
   if (nano_wait_thread_created) {
     if (nano_shared_aio != NULL)
@@ -290,8 +285,6 @@ SEXP rnng_thread_shutdown(void) {
     nng_cv_wake(nano_wait_cv);
     nng_mtx_unlock(nano_wait_mtx);
     nng_thread_destroy(nano_wait_thr);
-    nng_cv_free(nano_shared_cv);
-    nng_mtx_free(nano_shared_mtx);
     nng_cv_free(nano_wait_cv);
     nng_mtx_free(nano_wait_mtx);
   }
@@ -311,29 +304,25 @@ SEXP rnng_wait_thread_create(SEXP x) {
 
     int xc, signalled;
 
-    if (nano_wait_thread_created == 0) {
+    if (!nano_wait_thread_created) {
       if ((xc = nng_mtx_alloc(&nano_wait_mtx)))
         goto exitlevel1;
 
       if ((xc = nng_cv_alloc(&nano_wait_cv, nano_wait_mtx)))
         goto exitlevel2;
 
-      if ((xc = nng_mtx_alloc(&nano_shared_mtx)))
-        goto exitlevel3;
-
-      if ((xc = nng_cv_alloc(&nano_shared_cv, nano_shared_mtx)))
-        goto exitlevel4;
-
       if ((xc = nng_thread_create(&nano_wait_thr, rnng_wait_thread, NULL)))
-        goto exitlevel5;
+        goto exitlevel3;
 
       nano_wait_thread_created = 1;
     }
 
     nng_mtx_lock(nano_wait_mtx);
-    nano_shared_aio = aiop->aio;
-    nano_wait_condition = 1;
-    nng_cv_wake(nano_wait_cv);
+    if (nano_wait_condition == 0) {
+      nano_shared_aio = aiop->aio;
+      nano_wait_condition = 1;
+      nng_cv_wake(nano_wait_cv);
+    }
     nng_mtx_unlock(nano_wait_mtx);
 
     nng_time time = nng_clock();
@@ -341,15 +330,14 @@ SEXP rnng_wait_thread_create(SEXP x) {
     while (1) {
       time = time + 400;
       signalled = 1;
-      nng_mtx_lock(nano_shared_mtx);
-      while (nano_shared_condition == 0) {
-        if (nng_cv_until(nano_shared_cv, time) == NNG_ETIMEDOUT) {
+      nng_mtx_lock(nano_wait_mtx);
+      while (nano_wait_condition == 1) {
+        if (nng_cv_until(nano_wait_cv, time) == NNG_ETIMEDOUT) {
           signalled = 0;
           break;
         }
       }
-      nano_shared_condition = 0;
-      nng_mtx_unlock(nano_shared_mtx);
+      nng_mtx_unlock(nano_wait_mtx);
       if (signalled) break;
       if (!R_ToplevelExec(check_interrupt, NULL))
         nng_aio_stop(nano_shared_aio);
@@ -375,10 +363,6 @@ SEXP rnng_wait_thread_create(SEXP x) {
 
     return x;
 
-    exitlevel5:
-    nng_cv_free(nano_shared_cv);
-    exitlevel4:
-    nng_mtx_free(nano_shared_mtx);
     exitlevel3:
     nng_cv_free(nano_wait_cv);
     exitlevel2: