--- a/tests/tests.R
+++ b/tests/tests.R
@@ -666,6 +666,7 @@ test_type("integer", .Call(nanonext:::rnng_traverse_precious))
 if (Sys.info()[["sysname"]] == "Linux") {
   rm(list = ls())
   gc()
+  .Call(nanonext:::rnng_thread_shutdown)
   Sys.sleep(1L)
   .Call(nanonext:::rnng_fini)
   invisible()