This PR introduces a persistent wait thread.

- Launched first time it is needed (so not in mirai daemon processes for example, or other production pipelines not requiring interruptible waits)
- Now when a user interrupts, this will cancel the aio, causing it return `errorValue 20 | Operation canceled` rather than just returning the user to the console.
- The returned value may actually be more useful for logging purposes etc.

The performance gains, using the now-famous benchmark below - we're now practically at 1000x !!

Before:
``` r
library(mirai)
library(future)

daemons(8, dispatcher = "none", .compute = "none")
#> [1] 8
daemons(8, dispatcher = "thread", .compute = "thread")
#> [1] 8
daemons(8, dispatcher = "process", .compute = "process")
#> [1] 8
plan("multisession", workers = 8)

bench::mark(
  mirai(1, .compute = "none")[],
  mirai(1, .compute = "thread")[],
  mirai(1, .compute = "process")[],
  value(future(1)),
  relative = TRUE
)
#> # A tibble: 4 × 6
#>   expression                             min median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                           <dbl>  <dbl>     <dbl>     <dbl>    <dbl>
#> 1 "mirai(1, .compute = \"none\")[]"   1   e0   1         806.      7.86      Inf
#> 2 "mirai(1, .compute = \"thread\")[]" 1.20e0   1.15      712.      1         Inf
#> 3 "mirai(1, .compute = \"process\")[… 1.55e0   1.55      505.      1         NaN
#> 4 "value(future(1))"                  1.03e3 917.          1     278.        Inf
```

<sup>Created on 2024-11-01 with [reprex v2.1.1](https://reprex.tidyverse.org)</sup>

After:
``` r
library(mirai)
library(future)

daemons(8, dispatcher = "none", .compute = "none")
#> [1] 8
daemons(8, dispatcher = "thread", .compute = "thread")
#> [1] 8
daemons(8, dispatcher = "process", .compute = "process")
#> [1] 8
plan("multisession", workers = 8)

bench::mark(
  mirai(1, .compute = "none")[],
  mirai(1, .compute = "thread")[],
  mirai(1, .compute = "process")[],
  value(future(1)),
  relative = TRUE
)
#> # A tibble: 4 × 6
#>   expression                             min median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                           <dbl>  <dbl>     <dbl>     <dbl>    <dbl>
#> 1 "mirai(1, .compute = \"none\")[]"   1   e0   1         889.      7.86      Inf
#> 2 "mirai(1, .compute = \"thread\")[]" 1.18e0   1.15      778.      1         Inf
#> 3 "mirai(1, .compute = \"process\")[… 1.47e0   1.59      499.      1         NaN
#> 4 "value(future(1))"                  1.08e3 973.          1     278.        NaN
```

<sup>Created on 2024-11-01 with [reprex v2.1.1](https://reprex.tidyverse.org)</sup>