@wlandau 2024-11-04
> this would introduce a behavioural change for mirai / crew. 

Would you help me understand the differences in behavior, maybe with a toy example? Not sure I grasp it.

`crew` uses `nanonext::until_()` and `mirai::call_mirai_()`. At a glance, I think these are the only interruptible functions `crew` needs from `mirai`/`nanonext`.