@shikokuchuo 2024-11-08
This only affects your usage of `mirai::call_mirai_()`. If your crew functions are using that to wait, then just interrupt it with ctrl+c or Esc in Rstudio and you'll see that the mirai resolves to an `errorValue` now.