@shikokuchuo 2024-11-08
Actually Will, don't worry about this for now. I've greatly simplified the implementation and made it much more robust, and I think I can finesse it so it preserves much of the previous behaviour.